Break ILLiad SQL functions out to Database Object
